update and delete project (#4)

+++ src/api/project/project.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import {Injectable, NotFoundException} from '@nestjs/common';
 import {ProjectRepository} from "./schemas/project/project.repository";
 import {Project} from "./schemas/project/project.schema";
 import {CreateProjectRequestDto} from "./dto/create-project.request.dto";
@@ -23,12 +23,12 @@ export class ProjectService {
         return this.projectRepository.findAll();
     }
 
-    async updateProject(updateProps : Partial<Project>) : Promise<Project> {
-        if (updateProps._id == null) {
+    async updateProject(updateProps : Partial<Project>, projectId: string) : Promise<Project> {
+        if (projectId == null) {
             throw Error("Id for the project to update was undefined");
         }
-
-        return this.projectRepository.findOneAndUpdate({_id: updateProps._id}, updateProps);
+        await this.projectRepository.findOneAndUpdate({_id: projectId}, updateProps);
+        return this.projectRepository.findOne({_id: projectId});
     }
 
     async createProject(newProject: CreateProjectRequestDto): Promise<Project> {
@@ -36,9 +36,11 @@ export class ProjectService {
         return this.projectRepository.create(TemplateEngine.getTemplatedProject(project));
     }
 
-    async deleteProjectById(projectId: string) : Promise<DeleteProjectResponseDto> {
-        const results : {acknowledged: boolean, deletedCount: number} = await this.projectRepository.deleteProjectById(projectId);
+    async deleteProjectById(projectId: string, userId: string) : Promise<DeleteProjectResponseDto> {
         const deleteProjectResponse: DeleteProjectResponseDto = new DeleteProjectResponseDto();
+        const projectForDeletion = await this.getProjectById(projectId);
+        if(projectForDeletion.userId !== userId) throw new NotFoundException("Project not found for this user")
+        const results : {acknowledged: boolean, deletedCount: number} = await this.projectRepository.deleteProjectById(projectId);
         deleteProjectResponse.success = (results.deletedCount > 0);
         deleteProjectResponse.message = "Project deleted successfully";
         return deleteProjectResponse;
